Add project skeleton and integration tests.
The following have been set up.

An express server which renders the nunjucks templates in the components
directory. This is used for testing but we can use it for documentation
in the future.

Functional tests which use puppetteer to run a browser with the compiled
templates so we can run tests against them.